Check if replay exists when deleting

+++ server/src/routes/replays.ts
@@ -204,27 +204,34 @@ router.post('/', (req: Request, res: Response, next: Function): any => {
 router.delete('/:replayId', async (req, res) => {
     // Fetch replay
     const replay = await db.getReplayById(req.params.replayId);
-    const replayUser = await db.getUserById(replay.userRef);
 
-    // Check user
-    const { user } = req;
-    if (user === undefined || user.webId !== replayUser.webId) {
-        req.log.error('replaysRouter: Unauthenticated replay delete attempt');
-        res.status(401).send('Authentication required to delete replay.');
-        return;
-    }
+    if (replay) {
+        const replayUser = await db.getUserById(replay.userRef);
 
-    await db.deleteReplayById(replay._id);
+        // Check user
+        const { user } = req;
+        if (user === undefined || user.webId !== replayUser.webId) {
+            req.log.error('replaysRouter: Unauthenticated replay delete attempt');
+            res.status(401).send('Authentication required to delete replay.');
+            return;
+        }
 
-    try {
-        req.log.debug('replaysRouter: Deleted replay metadata, now deleting replay file');
-        await artefacts.deleteReplay(replay);
-    } catch (err) {
-        req.log.warn('replaysRouter: Failed to delete replay file, restoring metadata in database');
-        // If deletion failed, add the replay back into the DB
-        await db.saveReplayMetadata(replay);
+        await db.deleteReplayById(replay._id);
 
-        res.status(500).send('Failed to delete replay file.');
+        try {
+            req.log.debug('replaysRouter: Deleted replay metadata, now deleting replay file');
+            await artefacts.deleteReplay(replay);
+        } catch (err) {
+            req.log.warn('replaysRouter: Failed to delete replay file, restoring metadata in database');
+            // If deletion failed, add the replay back into the DB
+            await db.saveReplayMetadata(replay);
+
+            res.status(500).send('Failed to delete replay file.');
+            return;
+        }
+    } else {
+        req.log.error('replaysRouter: Replay not found');
+        res.status(404).send('Replay not found.');
         return;
     }
